Split vimrc to simple and full versions

+++ install.sh
@@ -1,6 +1,6 @@
-ln -nsf ~/dotfiles/.bashrc ~/.bashrc
-ln -nsf ~/dotfiles/.vimrc ~/.vimrc
+version=${1:-'simple'}
 
+ln -nsf ~/dotfiles/.bashrc ~/.bashrc
 sudo apt-get update && sudo apt-get install -y \
     curl \
     git \
@@ -20,12 +20,18 @@ echo 'set runtimepath^=~/.vim runtimepath+=~/.vim/after
 let &packpath=&runtimepath
 source ~/.vimrc' > ~/.config/nvim/init.vim
 
-# Install Plug (Vim plugin manager).
-curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
-	  https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
+if [ "$version" == "full" ]; then
+    ln -nsf ~/dotfiles/vim/full.vim ~/.vimrc
 
-# Install FZF (fuzzy finder on the terminal and used by a Vim plugin).
-rm -rf ~/.fzf
-git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf && ~/.fzf/install
+    # Install Plug (Vim plugin manager).
+    curl -fLo ~/.vim/autoload/plug.vim --create-dirs \
+          https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim
 
-pip3 install pynvim jedi
+    # Install FZF (fuzzy finder on the terminal and used by a Vim plugin).
+    rm -rf ~/.fzf
+    git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf && ~/.fzf/install
+
+    pip3 install pynvim jedi
+else
+    ln -nsf ~/dotfiles/vim/simple.vim ~/.vimrc
+fi
